python_phone_otp_service/requirements.txt
grpcio
grpcio-tools

twilio
requests

python-dotenv